+++ Calculator/Calculator/Model/CalculateItem/CalculatorItem.swift
@@ -1,0 +1,11 @@
+//
+//  CalculatorItem.swift
+//  Calculator
+//
+//  Created by 허건 on 2022/05/16.
+//
+import Foundation
+
+protocol CalculatorItem {
+    
+}
