+++ Calculator/Calculator/Model/CalculateItem/Double+Extension.swift
@@ -1,0 +1,10 @@
+//
+//  Double_Extension.swift
+//  Calculator
+//
+//  Created by 허건 on 2022/05/20.
+//
+
+extension Double {
+    
+}
