+++ Calculator/Calculator/Model/CalculatorQueue/ExpressionParser.swift
@@ -8,19 +8,21 @@
 enum ExpressionParser {
     
     /// 파서해서 Formular에 데이터 보내는 함수 변환?? 파싱한다..
-//    func parse(from input: String) -> Formula {
-//        let input: String
-//        
-//        return 
+//    static func parse(from input: String) -> Formula {
+//        componentsByOperators(from: input)
+//        let input: String = "3+3-3"
+//
+//
+//        return
 //    }
-    
-    /// 입력받은 요소들을 저장하는 함수
-//    private func componentsByOperators(from input: String) -> [String] {
+//
+//    /// Operators의 구성요소
+//    static private func componentsByOperators(from input: String) -> [String] {
 //        var arrays: [String] = []
 //        let input: String
-//        
+//        input.split(with: Operator.add.rawValue)
 //        arrays.append(input)
-//        
+//
 //        return arrays
 //    }
 }
