--- Calculator/Calculator/Model/Protocol/CalculatorItem.swift
@@ -1,23 +1,0 @@
-//
-//  CalculatorItem.swift
-//  Calculator
-//
-//  Created by 허건 on 2022/05/16.
-//
-import Foundation
-
-protocol CalculatorItem {
-    
-}
-
-extension Double {
-    
-}
-
-extension String {
-    func split(with target: Character) -> [String] {
-        var operators: [String] = []
-
-        return operators
-    }
-}
